Trait pallet_sunrise::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type SunrisePalletId: Get<PalletId>;
    type Cooldown: Get<Self::BlockNumber>;
    type MaximumRewardPerSwap: Get<Balance>;
    type LeftoverSwapRebates: Get<FixedU128>;
    type Security: SecurityExt<Self::AccountId, Self::BlockNumber>;
    type CurrencyTidefi: Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Events

type SunrisePalletId: Get<PalletId>

Pallet ID

type Cooldown: Get<Self::BlockNumber>

Number of blocks to wait before allowing users to claim their sunrise rewards, after an era is completed.

type MaximumRewardPerSwap: Get<Balance>

Maximum rewards for sunrise pool before the applied rewards. Example, 10_000_000_000_000_000 with a 125% rewards, the maximum applied will be 12_500_000_000_000_000 TDFY

type LeftoverSwapRebates: Get<FixedU128>

For each tier, leftover funds will be allocated to to this tier

type Security: SecurityExt<Self::AccountId, Self::BlockNumber>

Security traits

type CurrencyTidefi: Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance>

Tidechain currency wrapper